stage1/pyproject.toml
[project]
name = "stage1"
version = "0.3.0"
description = "Description"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110.0",
    "gradio>=5.22.0",
    "httpx>=0.28.1",
    "mcp-server-fetch>=2025.1.17",
    "openai>=2.0.0,<3",
    "openai-agents>=0.4.0,<1",
    "playwright>=1.51.0",
    "psutil>=7.0.0",
    "pydantic>=2.10.4,<2.12",
    "pypdf>=5.4.0",
    "python-dotenv>=1.0.1",
    "python-json-logger>=4.0.0",
    "semantic-kernel>=1.25.0",
    "sendgrid>=6.11.0",
    "smithery>=0.1.0",
    "uvicorn>=0.34.0",
    # Server authoring (resources + tools)
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.10.0",
    "pytest>=8.0.2",
    "pytest-asyncio>=0.23.5",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore:websockets\\.legacy is deprecated:DeprecationWarning",
    "ignore:websockets\\.server\\.WebSocketServerProtocol is deprecated:DeprecationWarning",
]

[tool.ruff]
line-length = 100

[build-system]
requires = ["setuptools>=78.1.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
